[KNI][REVERT][CARRY] nrt: cache: add support for dedicated informer"
This reverts commit ba793a9e0cd5f98189c9df5d2c81af70e8e4d38d.
