+++ apis/config/v1/defaults_test.go
@@ -205,7 +205,6 @@ func TestSchedulingDefaults(t *testing.T) {
 				Cache: &NodeResourceTopologyCache{
 					ForeignPodsDetect: &defaultForeignPodsDetect,
 					ResyncMethod:      &defaultResyncMethod,
-					InformerMode:      &defaultInformerMode,
 				},
 			},
 		},
