+++ apis/config/v1beta3/zz_generated.deepcopy.go
@@ -220,11 +220,6 @@ func (in *NodeResourceTopologyCache) DeepCopyInto(out *NodeResourceTopologyCache
 		*out = new(CacheResyncMethod)
 		**out = **in
 	}
-	if in.InformerMode != nil {
-		in, out := &in.InformerMode, &out.InformerMode
-		*out = new(CacheInformerMode)
-		**out = **in
-	}
 	return
 }
