+++ apis/config/zz_generated.deepcopy.go
@@ -170,11 +170,6 @@ func (in *NodeResourceTopologyCache) DeepCopyInto(out *NodeResourceTopologyCache
 		*out = new(CacheResyncMethod)
 		**out = **in
 	}
-	if in.InformerMode != nil {
-		in, out := &in.InformerMode, &out.InformerMode
-		*out = new(CacheInformerMode)
-		**out = **in
-	}
 	return
 }
